Added Helper to Add Example Puzzles
