+++ package.json
@@ -26,7 +26,8 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "add-puzzles": "ts-node --compilerOptions '{\"module\": \"commonjs\"}' scripts/addPuzzles.ts"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -39,6 +40,7 @@
   ],
   "devDependencies": {
     "prettier": "^1.15.2",
-    "surge": "^0.20.1"
+    "surge": "^0.20.1",
+    "ts-node": "^7.0.1"
   }
 }
