+++ src/database.ts
@@ -1,3 +1,6 @@
+import firebase from "firebase";
+import { firestore } from "./firebase";
+
 export interface PuzzleWithId extends PuzzleData {
   id: string;
 }
@@ -11,3 +14,32 @@ export interface Database {
   getPuzzle(id: string): Promise<PuzzleWithId | null>;
   newPuzzle(data: PuzzleData): Promise<PuzzleWithId>;
 }
+
+class FirebaseDatabase implements Database {
+  private db: firebase.firestore.Firestore;
+
+  constructor(firestore: firebase.firestore.Firestore) {
+    this.db = firestore;
+  }
+
+  async getPuzzle(id: string): Promise<PuzzleWithId | null> {
+    const doc = await this.db
+      .collection("puzzles")
+      .doc(id)
+      .get();
+
+    const data = doc.data();
+    if (!data) {
+      return null;
+    }
+
+    return { id: doc.id, ...(data as PuzzleData) };
+  }
+
+  async newPuzzle(data: PuzzleData): Promise<PuzzleWithId> {
+    const docRef = await this.db.collection("puzzles").add(data);
+    return { id: docRef.id, ...data };
+  }
+}
+
+export const database = new FirebaseDatabase(firestore);
